pyproject.toml
[project]
name = "template"
version = "0.1.0"
description = "Agente de AI como plantilla para desarrollo con agentes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.55",
    "google-adk>=1.16.0",
    "litellm>=1.78.5",
]
